requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.2.0
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
dnspython==2.3.0
ecdsa==0.19.0
email-validator==2.0.0.post2
exceptiongroup==1.2.2
fastapi==0.103.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.7
importlib-metadata==6.7.0
iniconfig==2.0.0
motor==3.4.0
packaging==24.0
passlib==1.7.4
pluggy==1.2.0
pyasn1==0.5.1
pydantic==2.5.3
pydantic-core==2.14.6
PyJWT==2.8.0
pymongo==4.7.3
pytest==7.4.4
python-dotenv==0.21.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
tomli==2.0.1
typing-extensions==4.7.1
uvicorn==0.22.0
zipp==3.15.0
